refactor the delete favourite fetch
